driver*,prop,rcu,xb,tests: Add support for race detection in LKMM
This commit adds the necessary infrastructure to support races
under LKMM. It also adds a handful of test cases that contain
plain accesses.

However, some race-related definitions are not quite right yet.
For instance, when calculating pre-bounds, only addr dependencies
are used instead of the full ppo. In addition, a (strange) plain
coherence axiom is used that is not yet implemented in GenMC.
Finally, many other differences with herd and heuristics used by
GenMC need to be vetted.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>