config,driver: Change scheduling policy and add CLI option
This commit changes the default scheduling policy from
"left-to-right" to "writes-first". The latter tries to first
execute threads the next instruction of which is not a load.
This improves the performance of the tool in a variety of
different benchmarks.

In addition, the commit adds a CLI option that allows users to
choose the scheduling policy used by the tool. This renders the
old -randomize-schedule option obsolete, as it is now incorporated
in the available scheduling policies.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>