driver: Make the fallback scheduling policy of WF to be LTR
For some benchmarks, employing writes-first actually decreases
performance. For instance, cinc(5) becomes 1.5x slower when using
WF (and choosing the rightmost thread as a fallback option).

This commit attempts to ameliorate the performance of the tool
when WF fails, by employing LTR as a fallback strategy. In general,
the test suite becomes slower with this change (270sec -> 278sec),
but this is attributed to a slowdown in LAPOR and in test cases
in the "wrong" folder. I think what these have in common is the
fib_bench benchmark (same as triangular-1 in LAPOR), which seemed
to benefit a lot from an LTR policy. Still just this benchmark
probably does not justify the overall usage of LTR, as the
average time in other categories decreases.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>