depexecgraph,tests: Fix nasty bug w/ pporf views and rfis
This commit fixes a bug related both to how dependencies are
calculated and to how prefixes are saved (in a DepExecutionGraph).

The bug is easily manifested in LB+RA. Specifically, LKMM dictates
that only "dep; [Marked]; rfi" is preserved to reads (as opposed
to "rfi U dep"). Thus, in the first thread, there is no dependency
from Rx to Wyr_1, and the Wx1 of the second thread can revisit
the Rx of the first thread in the execution where the Ry of the
second thread reads from Wyr_1. However, in said revisit, the prefix
we are saving needs to include Ws1, despite the fact that Ws1
is not in the pporf prefix of Wx1 (only the Rs part is). Thus,
we need to include rfis when saving prefixes, even if these rfis
are not part of the write's prefix.

Note that we cannot just include rfis in the pporf views (or
conditionally include them in the pporf views), as other test
cases break in that case (e.g., MP+RA).

[This fix needs to be migrated to IMM, along with the relevant
fix in the calculation of dependencies and views.]

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>