driver,interp: Keep just 1 symmetric thread and fix symmetry check
This commit optimizes the number of symmetric threads kept per
thread by only keeping the thread with the greatest ID as the
symmetric one (among all threads that are similar).

It also makes sure that the parents of two threads are equal
before declaring them as symmetric.

TODO: The commit does _not_ check that there are not memory
accesses between fork events that create symmetric threads.
This is because currently all fork events generate a memory
access event corresponding to the store of a threads TID to a
user-specified location.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>