include,interp: Change the way thread creation is modeled
Until now, pthread_create() used to directly boil down to
__VERIFIER_thread_create(). However, this was leading to the
creation of memory accesses between different thread creations,
accesses that were corresponding to the store of the TID of the
newly created thread. The location of the TID was not promoted
to a register by mem2reg since its address was passed to
pthread_create(), thus leading to extra alloca and store
instructions.

This commit alleviates this problem by modeling pthread_create()
as a macro that simply writes the TID returned by
__VERIFIER_thread_create() to a user-specified location (after
__VERIFIER_thread_create() returns). This yields a 3-4%
performance benefit in the correct test cases of the test suite.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>