interp,driver: Categorize blocked executions
This commit introduces a new type "BlockageType" that helps
in categorizing the different reasons a thread may block. This
is useful for statistics or for liveness checks (see subsequent
commits).

CHECK:
I am not sure whether the type "LockRel" is necessary.
Now it is used when blocking executions where a thread did not
release a lock in LAPOR. But maybe this can be subsumed by
two different types "Unroll" + "User", if we decide to not count
executions where a loop is unrolled more than the allowed limit.

Check again all calls to block() and make sure that:
    1) The argument makes sense
    2) The call itself makes sense

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>