config,interp,driver,spinassume: Check for liveness violations
This commit adds support for reporting some liveness violations
involving spinloops. To do so, each assume statement coming from
a spinloop transformation is annotated accordingly, and a liveness
violation is reported when: (a) the relevant switch has been
enabled, and (b) all reads associated with the spinloops' assumes
are reading a non-maximal write.

TODO: Currently, this assumes that consistency checks at the end
of each execution are enabled, in order to be able to check for
liveness violations when LAPOR is enabled (and hence coherence
requires a fixpoint calculation). However, checking consistency
at each execution should be unnecessary for the non-LAPOR case.
This should be subsequently fixed.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>