co,driver: Fix the way co-maximal writes are determined
This commit fixes the way maximal writes in coherence are detected,
by consulting the fixpoint calculation if it is necessary (e.g.,
if LAPOR is enabled), and by asking the coherence calculator
otherwise.

(Perhaps this can be further improved by making EG a state machine,
and always consult the calculators to determine whether a store
is maximal. The calculators would in turn check the status of the
EG and either locally calculate coherence, or check the fixpoint
result.)

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>