driver: Complain when trying to read from uninitialized memory
This commit makes GenMC complain whenever a read is trying to
access a dynamically allocated variable that is not initialized.

Such a check should occur in two places: whenever a read is
added (with the read's initial rf), and when the rf of a read
is changed. Although not visible with the current test cases
and the current way rfs are selected (see stack-uninit2), it can
be the case that the first rf that is selected corresponds to
a write other than INIT, and only after restriction does the
read in question read from INIT.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>