pb,rcu-link,tests: Add the ar* part at the end of pb (ad-hoc)
This commit makes sure that the ar* part at the end of pb is
added where necessary. For most cases, since xb = (ar U pb U rb)+,
the trailing ar part of pb is added by xb. However, for rcu-link,
the ar part of pb is necessary. Thus, instead of always calculating
it in pb, we calculated only when calculating rcu-link.

A test case demonstrating the problem is added as well.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>