conf,interp,driver: Fix memory leak caused by interp ownership
This commit fixes a long-standing issue that was causing a (minor)
memory leak. The problem was that the interpreter was not freed
by the driver (as such an action would cause a double-free error)
due to the module not being exclusively owned by either the driver
or the interpreter.

Now that support for LLVM-3.5 has been dropped, a fix was easy to
implement, and full ownership of the module could be handed over
to the interpreter.

Currently, the ownership chain is as follows:

  * Driver -> EE -> Module

where "->" means "owns". In addition, the driver has no longer
a reference to the underlying module, since all necessary info
can be obtained through the interpreter.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>