driver: Do not check for uninitialized memory for internal variables
Assume that we know what we are doing and do not check for
uninitialized memory locations for internal variables. Otherwise,
the EE might try to replay an execution that contains an event
(part of a syscall) that is reading from uninitialized memory
and get stuck in an infinite loop (because if the syscall blocks
the EE's state gets reset to the beginning of a syscall; this
happens to model failed lock acquisitions).

Also, for FS test cases, it is common to, e.g., do CASes on variables
that are not not stored in the static array (for better naming info),
and these will trigger this error check. This needs to be fixed
from a subsequent commit that will rewrite the memory allocator.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>