eventlabel,execgraph,driver*,tests: Fix bug with trylock()
This commit fixes a long-standing bug where the tool would
complain that an unlock unlocks a lock locked by a different
thread, even though that was not the case. This could happen
in test cases w/ trylock operations, because these operations
were not properly treated as locks.

(This needs to be migrated to dev.)

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>