include,pass*,interp*: Add special marker for spinloop halting
This commit adds a special function that acts as a marker for
the end of spinloops so that we don't use assume()s for anything
else other than user-specified assumptions. The new function is
called __VERIFIER_spin_end(), and can also be called by the users
(along with __VERIFIER_spin_start()) to denote the end/start of
spinloops that the tool does not automatically transform into
assume() statements.

(Test cases pending.)

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>