interp,tests: Fix Persevere bug w/ fsync() and external synchronization
This commit fixes a bug that occurs when there is an fsync() after
a read() that reads externally. An fsync() after a read() of the same
file will make sure that all (disk) operations for said file will
have persisted when the fsync() returns. However, due to the fact
that writes used to first update the memory size and then the
disk size, the fsync() would not be able to ensure that the size update
before the read() would have also persisted.

This commit fixes this issue by having size updates first update
the disk size, and then the memory size (in contrast to what the
actual implementation does, but that should not matter).

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>