depexecgraph,immdriver: Fix rfi bug in ppo
This commit fixes a bug in IMM's calculation of ppo where rfi
edges would unconditionally be included in ppo. This bug would
manifest in test cases like rfi-preserved where the tool would
explore one less execution than what it should.

The fix is a migration of the following LKMM fixes:

8e2f2abe depexecgraph,tests: Fix nasty bug w/ pporf views and rfis
013baaf7 lkmmdriver: Fix ppo definition for LKMM (rfi;dep problem)

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>