immdriver,tests: Fix release view calculation in ppo
This commit repairs the calculation of release stores to include
the pporf view of previous release stores. Without the fix GenMC
could crash in test cases like rel-pporf, where the second
release store in runA does not include the pporf prefix of the
first release store. Therefore the store "x=1" of runB is not
included in G|_{a}Uprefix(a) on the revisit of R(y) in runB
which reads from the second release store.

Relevant Github issue:

    https://github.com/MPI-SWS/genmc/issues/1

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>