lkmm: Bring dependency calculation up to date (merge pending)
This commit updates dependency calculation to take into account
recent fixes in dev branch (which in turn were the result of
migrating fixes from lkmm).

Note: A merge is pending in order to account for some changes
in depexecgraph; a rebase would not do as it would render some
of the commits in this branch (containing ppo fixes)
weird/useless.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>