pass*,eventlabel,interp,driver*: Towards optimizing FAI spinloops
This commit is the first step towards optimizing FAI spinloops
into assume()-like statements. It introduces the necessary
infrastructure in the graph and the driver, as well as a
mechanism for passing information from the interpreter to the
driver regarding potential FAI spinloops.

The next step is to properly recognize such spinloops, as
currently only a "useless" form of such loops is recognized.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>