callinfo,llvm*,spinassume: Don't assume calls have side effects
In spin-assume, CallInsts were considered to have side effects
but that is not always true. A function might not touch memory
at all, in which case calling it should not be considere to
have side effects.

This commit introduces a new pass that scans the code for
side-effect-free functions, so that this info can be used by
subsequent passes (e.g., spin-assume). One example where
this is important is 'treiber-stack', where the spin-assume
transformation now succeeds.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>