annot: One effect-free path is enough for a load to be annotatable
Even if there is only one path from a load to an assume with no
side-effects, it is worth it to annotate the load. In principle,
we don't even need to check for side-effects, but checking filters
out some loads that will have blocking conditions that will
always be false.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>