driver*,tests: Bug fixes for nested thread creation
This commit fixes a couple of bugs having to do with nested
thread creation. It ensures that thread begins properly
synchronize with their parent creates and makes sure that
threads the parent create of which has been replaced with a
different event will not be scheduled. See infr/nested-create
for an example that now works.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>