cohcalc*: Add a cache for recent calculations (WB)
This commit adds a cache that stores info from the most recent
coherence calculation (for a particular memory location). In
cases where we make repeated queries for coherence information
(e.g., when filtering annotation values) this cache can improve
the running time by a large margin.

Currently info about only one memory location (and a single
calculation) are stored, but it should be possible to extend the
cache to hold more detailed information.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>