error: Refine debug-printing infrastructure
This commit introduces a format() function that can be used
to transform STL (and LLVM) containers and std::pair<>s into
std::strings. In fact, it should "just work" for any container
offering begin() and end() iterators.)

Note: Overloading operator<< for llvm::raw_ostream does not seem
such a good idea, as there are a few container-like classes in
LLVM that do not have the required iterators.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>