execgraph,cohcalc*: Fix bug, cons checks for many dynamic allocs
This commit fixes a bug (stack overflow) occurring when performing
consistency checks in programs with an excessive amount of
allocations/deallocations.

GenMC assumed that allocation/deallocation happened rarely (e.g.,
at the beginning of each program by main()). So, it did not check
whether a location was "alive" in a particular execution when
performing consistency checks. However, if memory was allocated
anew at each execution, a lot of new memory locations was created,
which rendered (coherence) consistency checks costly. For WB,
calculating linearizations of locations that were not alive
killed the tool's performance.

This commit fixes the above by only keeping locations that are
alive in each execution, and completely forgetting the existence
of the rest.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>