*: Add built-in support for barriers
This commit adds support for barriers (e.g., pthread_barrier).
It also optimizes their handling and doesn't explore N! executions
for the RMW chains that make up a barrier rendezvous, but rather
just 1 (provided that not more than N threads rendezvous, where
N is the value the barrier is initialized with).

Error checks specific to barriers and cleanups are pending.

(This commit is a rewrite of old code written by Xiaowei.)

Co-authored-by: Xiaowei Ren <renxiaowei66@gmail.com>
Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>