driver: Fix bug in isCoMaximal()
This function would only consult the calculator's cache if
full consistency checks should not be triggered. However, that
would lead to erroneously _not_ check the cache in cases like
LAPOR, even if the user had explicitly asked the cache to be
queried.

This commit makes isCoMaximal() query the cache if the user
explicitly specifies it, irrespective of whether consistency
checks will be performed or not.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>