driver: Fix load annotation bug (SAVer)
Value filter should only take place when a read is added, and
not when revisiting. As an example of why, see szymanski with
consistency checks enabled at each step.

This commit addresses this issue. (Now loads do not have to
carry annotations in their labels, but this commit does not
change the labels.)

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>