eventlabel,driver: Fix bug w/ in-place revisiting
One of the conditions for actually performing a revisit (w,r)
is that r has not been already revisited. However, this leads
to a completeness issue in programs with locks and the in-place
revisiting optimization. Consider the following:

  lock()   || lock()   || lock()
  unlock() || unlock() || unlock()

Suppose we have arrived at the following execution (w/ temporary
functionality violation), with event (2,1) being added after
(1,1) but before (1,2):

  L(0,0) L(1,2) L(0,0)
  U      U

The fact that (2,1) was added after (1,1) means that (2,1) was
revisited in place by (1,2). In turn, the in-place rev implies that
(3,1) will not revisit (1,1) due to (2,1) "having already been
revisited".

The problem stems from the fact that we will not explore the
execution where (2,1) reads from (1,1) due to the in place
revisit, and thus (3,1) will never be able to revisit (1,1).

This commit solves this problem by relaxing the condition about
"reads that have already been revisited" to exclude in-place
revisits.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>