scripts,tests: Redesign test suite argument parsing
There used to be separate files for the arguments taken by GenMC,
clang, as well as the unroll value. While this structure was
somewhat helpful in determining how a test case can be manually
run (by inspecting which files existed), it was also somewhat
restricting, as it wouldn't allow e.g., the same test to be
run with multiple different arguments for GenMC.

This commit rectifies this by having a single args.$model.$coh.in
file that contains both the arguments to be taken by genmc, as
well as the arguments to be taken by clang, with the two
separated by '|'.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>