spinassume,eventlabel,*: Fix SAVer bug w/ multiple loop calls
SAVer used the previous spin-start event to determine whether
a particular loop iteration makes progress or not. This is
problematic in cases where a loop is called via a function.

This commit addresses this issue by explicitly marking the
entrance of loops that for which DSA applies. Thus, SAVer can
now check for spin-start events after the marking denoting the
beginning of a particular loop.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>