config, tests: Disable LAPOR under LKMM
LKMM's ppo calculation leads to some executions being missed
with LAPOR (e.g., binc-locks). This is due to writes not being
ppo-after release fences (and hence LAPOR unlocks) in LKMM.

A possible fix would be the following: whever a LAPOR lock is
included in the prefix of a revisiting write, its matching unlock
should also be included, in order not to have empty labels in
the new graph.

Apart from the above, however, more fixes (e.g., in PROPCalculator
would also be necessary).

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>