interp,driver: Make back-rev explorations recursive (TODOs)
This commit makes the exploration of backward revisits to happen
recursively so that the memory consumption remains polynomial.
It also adds "state" structures that are going to be useful for
the parallellization.

Finally, it fixes some very _nasty_ bugs that occured while
making the exploration recursive, but proper fixes for these
will have to be pushed later.

TODOs:
    - grep FIXME, IMM, NOTE
    - Making the exploration recursive for IMM is going to be
      hard, especially due to bugs like the one that invalidated
      references within the EE. Try and make that on a separate
      branch, and then merge the RC11 and IMM branches together.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>