moduleinfo,*: Decouple essential info from a particular Module
This commit decouples information like annotated loads or variable
names from particular pointers within a module. Instead,
ModuleInfos now keep a mapping of IDs -> Values and associate
useful information to IDs (and not to Values), so that they can
be safely and easily copied to different ModuleInfos, presumably
in different contexts too.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>