sexpr*,annot*,pass*,interp: Make SExpr independent of LLVM types (1)
This commit rewrites a large part of the SExpr infrastructure so
that it does not utilize LLVM info like llvm::Type and llvm::Value.
This is necessary to be able to use annotations e.g., with
multiple GenMC threads.

That said, this commit is only the first step toward that direction.
We also have to make sure our "address translation" mechanism
assigns IDs to function arguments, so that these can be used in
SExpr across different threads.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>