scripts,config,tests: Allow -check-liveness only under MO
Allowing liveness checks under WB is complicated: even if
a loop seems dead (due to reads reading wb-maximal), the
execution with these wb-maximal values might not be consistent.
We have to check that the extension of wb for which the loop
loads are coherence maximal is consistent, and then determine
the liveness violation.

OTOH, checking for coherence maximality with WB probably beats
the purpose of using wb in the first place, so let's disable
the liveness checks under wb until someone complains about it.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>