driver,tests: Fix liveness checks under DSA
Under DSA, a spinloop is blocked at the beginning of the (n+1)-th
execution instead of at the end of the n-th (n \in N+). This
creates an issue for the liveness checking mechanism, as it
would only check events up to the last spin-start label.

This commit corrects this issue. Fixes Github #23.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>