interp,driver,*: Restore FS test cases (+ fix necessary bugs)
This commit performs the necessary changes in order to restore
the filesystem test cases for Trust.

The main reason these broke in the first place was the introduction
of the address translation mechanism. In order to have a simple
mapping from the directory inode's contents to the addresses of the
inodes of the contained files, we used to keep a map
FILENAME -> INODE_ADDR, where FILENAME was a const char *.
The allowed map keys were collected statically using a pass.
When the address translation mechanism was introduced, however,
the statically-collected keys would differ from the ones used
dynamically (since we use mock addresses during runtime). Worse
yet, we could not translate back to the statically collected ones:
the static filename basically was the value of the respective
global variable (which contained a const char *), and the only
mapping we kept was MOCK GLOBAL -> CONTENTS, and not
CONTENTS -> MOCK GLOBAL (as the latter was unnecessary).
Finally, having FILENAME being a const char * was not safe
as far as concurrency is concerned, since we want to be able
to pass this map to different threads.

To deal with the above issue, we change the type of FILENAME
to be a string. That way, we can both collect the allowed
filenames statically, and pass the map to different threads.

Along the way of fixing the above, the following were also
fixed:

    1) The local state of the interpreter was updated to include
    both the status of the interpreter, as well as some FS
    info that change dynamically are need to be saved as well
    (e.g., FD table).
    2) Some bugs introduced when replacing GVs with SVals.
    3) ModuleInfo should not be changed dynamically now.
    Components that change dynamically have no place being there.

FS tests also revealed a very nasty bug that will be fixed in
a subsequent commit.

*** Lesson learned: NEVER, EVER, EVER break the test suite again ***

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>