interp: Fix nasty bug w/ global pointers to globals
This commit fixes a bug uncovered when restoring FS
test cases (see parent commit), for nano-backup-proc.
The problem is caused by the address translation mechanism
in cases where a global maintains the address of another
global.

During the exploration, we use mock addresses for globals:
instead of having globals vars directly mapped to the address
of their contents (bc we have different copies of each global
per thread), we have them mapped to some mock address (same per
thread), and we maintain a (per-thread) mapping between the
mock and the actual contents to read the value when necessary:

(old)
    EE:
       @x --> [42]
       @y --> [@x]

(cur)
    EE:
       @x ------> x'
       @y -> y'   ||
             ||   ||
            [@x] [42]

The problem is that if a global was initialized with the
address of another global, it has the actual address
and not the (updated) mock address, as can be seen above
([@x] should have been [x']).

This commit fixes the problem by properly re-initializing
all global variables after the address translation
has been established.

---

This would have been much easier to diagnose if we could
immediately pinpoint that it is due to the introduction of
the address translation mechanism (and not due to SVals or
something else).

*** Once again: NEVER, EVER break the test suite again ***

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>