execgraph,driver: Restore LAPOR w/ TruSt (TODOs)
This commit makes all necessary changes to restore LAPOR for
TruSt. It also fixes a (pre-existing) bug with thread prioritization
that prevented TruSt from working (e.g., in complex-lock). The
issue was that, during thread prioritization, the driver would
not stop executing a prioritized thread even after the prioritized
CS was over.

Note that the commit does not restore LAPOR's tests as some of
them rely on SAVer anyway.

TODOs:

    - After SAVer is fixed, restore LAPOR tests
    - Fix the fixpoint infrastructure so we can remove the ugly
      consistency checks from all over the place

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>