eventlabel,interp,driver,atype: Account for the type of mem accesses
This commit introduces the type AType that is stored in
EventLabels and can be used to track the type of accesses
performed by the program. While this is not very useful in the
presence of ASize, it can be used to prettify the printing of
graphs: printing everything as unsigned integers is not very
helpful.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>