eventlabel,driver*,interp: Make helping CASes dummy labels
Helping CASes have basically no reason to exist as actual RMW
labels so long as the annotation condition is satisfied: all
stores visible to the helped CAS need to be visible to the
helping CAS as wll. This commit transforms the helping CAS label
into a dummy one, while maintaining all the info we kept so far.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>