include,interp,elimannots: Change modeling of non-value-returning RMWs
Until now, non-value-returning RMWs were represented as calls to
intrinsic functions. This had the advantage that we did not have
to annotate the instructions at the IR level, and that the function
call itself would not return any value.

The above modeling, however, makes it difficult to deal with

    1) CASes, as these should return the old value at the address
    of their second argument (which we can only model as an actual
    store, and not as a store to a register)
    2) different datatypes, as having polymorphic function leads
    to inherently ugly solutions in C

This commit remodels the representation of non-value-returning
RMWs by using a marker just before the RMW instruction that should
be treated as non-value-returning, and then having a dedicated
pass actually annotate the instruction.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>