interp,driver,*: Use value_ptr<>s for annotations; restore SAVer
This commit replaces the usage of shared_ptr<>s in ReadLabels
with value_ptr<>s so that these annotations can be safely used
in a concurrent environment.

It is possible that we can also use plain shared_ptr<>s since
only the maximal load's annotation will be evaluated at all times,
but let's try to be safe for now. Copying annotations shouldn't
be expensive anyway since (a) not all loads are annotated and (b)
annotations are usually small.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>