elimcasphis: Add a pass that eliminates certain CAS PHIs
Whenever clang produces a CAS instruction, that is accompanied
by a PHI node the value of which depends on whether the CAS
succeeded or not.

In certain cases, however, this PHI node can be eliminated.
More precisely, if:

    1) the Φ returns the expected value of the CAS whenever the
    CAS succeeds, and the value read when the CAS fails, and
    2) the block coming from the failed CAS branch is empty and
    unconditionally jumps to the Φ,

then we can simply remove the failed-CAS block, and replace all
uses of the Φ with the value read by the CAS.

This transformation aids our bisimilarity pass in cases like RDCSS.

As some extra help, this commit also enables LLVM's JumpThreading.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>