unroll: Revamp loop unrolling pass (+ no dependence on mem2reg)
The loop unrolling pass had a couple of issues:

    1) It required another pass of mem2reg afterwards, as it
    inserted an alloca, as well as stores and loads to the new
    memory location for the bound counter.

    2) It did not work properly if there were multiple latches.
    For instance, consider the following program:

        B0:
            a = Φ ((v1, B1), (v2, B2))
            ...

    The loop unroll pass would set the "diverge" as B1's and B2's
    successor. That way, however, we end up with the following
    B0

        B0:
            a = Φ ((v1, diverge), (v2, diverge))

    which is clearly wrong.

This commit rectifies these issues by making the unroll counter
effectively a Φ node. The counter value is subsequently checked
by (the modified) __VERIFIER_end_loop() and, if the loop is
free to execute, decreased. (All the above happen in the header
of the loop. The CFG remains intact.)

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>