dce: Add an SMC-safe unused code elimination pass
This commit adds a simple DCE pass that gets rids of some
instructions that have no uses. (Initially I thought that this
pass would be helpful for jump threading, but it turned out that
this is not the case. Still, it could make the interpereter a bit
faster so, why not?)

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>