instannot,jumpthreading: Improve loop jump threading pass
This commit improves our loop jump threading pass so that we can
perform DSA in functions like ms-queue's dequeue().

Idea: if we have a Φ0 in the header such that:

    1) Φ0 has a constant incoming value from the (unique) loop
    predecessor that makes the loop jump into its body

    2) Φ0 is either not used in the body or is only used in other
    Φ nodes which in turn are only used in Φ0

then we can perform jump threading and always jump to the loop
body from the predecessor, provided that we replace all uses of
Φ0 in the loop body with its constant incoming value.

In effect, this captures cases like the following:

    success = false;
    while (!success) {
        if (...) {
            // 'success' is not modified.
            (continue)
        } else {
            success = ...;
            (continue)
        }
    }

where an extra Φ (Φ0, ...) ψαν be produced after the conditional
that joins the values of the two branches. In such a case it is
safe to do jump-threading as we know that, if the second Φ gets
the value of the Φ0, then that value will be 'false'; otherwise,
we would have exited the loop (see spin-cas-onebranch).

This reduces the number of executions explored in ms-queue.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>