execgraph,driver: Redesign fixpoint infrastructure
This commit makes the fixpoint infrastructure of the execution
graph "stateful". Instead of having to call isConsistent()
before we need some more complex information from the execution
graph, the graph now has some internal state, calculates the
fixpoint on demand, and avoids doing it again once it has been
already calculated.

An API that may trigger the calculation has been defined which
includes:

    - isConsistent()
    - isHbBefore()
    - isCoMaximal()

More cleanups to ensure that the driver uses this API only are
required.

This is the third (and final) step toward the redesign of
the fixpoint infrastructure.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>