execgraph,calc*,driver*: Get rid of ExecutionGraph::getHbBefore()
This is just confusing. With the new fixpoint infra, clients
should either call ExecutionGraph::isHbBefore() or get an
approximation from the views.

(isHbBefore() itself can also provide an approximation, but in
places we already know an approximation will suffice, we can skip
the extra indirections. Perhaps it'll be a good idea to abstract
the whole view manipulation story away from the driver.)

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>