labeliter: Introduce fancier iterator for execution graphs
A very common usage pattern on EGs is iterating over their labels.
The way this was done, however, was often error-prone, since we
had to carefully calculate the indices we are calculating over,
and then get the relevant labels.

This commit makes iterating as simple as:

    for (const auto *lab : labels(G))
        ;

where G is an ExecutionGraph ref/ptr.

The implementation is a bit ugly and assumes that the graph
contains unique_ptr<>s, but hey, it works. It might also be
the case that using the new iterator is slightly faster, since
we avoid calling getEventLabel() which is not inline (yet).

(A couple of usage examples have been added as part of this commit.)

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>