execgraph,wbcalc,tests: Fix bug with BAM-TruSt interaction
This commit fixes a (long-standing) issue involving BAM and TruSt,
as some conditions in the maximality checks of TruSt have to be
relaxed in order for BAM to work.

As an example, TruSt checks whether revisiting a read will cause
the rf of some read to "dangle", and if that is the case it aborts
the revisit, with the rational that the revisit could have been
performed in some other execution, where there would be no
dangling rf. For BAM, however, we do not want to abort the revisit
when the dangling rf is the rf of a barrier operations, as there
will not be an alternative execution where the rf will not dangle:
BAM revisits reads in-place, and only explores one execution
as far as the interaction among barriers is concerned.

This commit makes sure that the relaxed maximality checks take
place only when BAM is actually enabled.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>