execgraph: This is why I hate user-defined copy constructors
Since I am dumb and forget to update the cc when I modify the
class fields, I deserve the free ticket to the UB-land.

Perhaps, instead of using getCopyUpTo(), it would make sense to
use value_ptr<>s for the labels of the graph, then use the default
copy constructor and simply cut the copy instead of using
getCopyUpTo()?

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>