interp,enumAPI: Fix dependencies for barriers + recursive calls
This commit ensures that we do take dependencies into account
when diving into pthread_barrier_*() calls (this would have been
fixed by a rebase anyway), but also ensures that the dependencies
are reset before calling driver->visitLoad() during a barrier_wait().

The reason for the latter is that if we initiate a recursive
exploration from the RMW of barrier_wait(), the dependencies might
be completely messed up when we call visitLoad().

The commit _cheats_ and sets them to nullptr, since the previous
accesses are RA anyway, their dependencies are correctly set,
and the load produced by visitLoad() will thus correctly take
any addr and ctrl dependencies into account.

TODO:
    - Properly fix this by redesigning the API between the
      driver and the EE

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>