depexecgraph,mocalc,immdriver: Fix IMM-MO with TruSt (?)
This commit fixes some bugs with forced reads/writes in IMM-MO
under TruSt. To that end, it relaxes the definition of a "prefix"
so that it contains not only events that are pporf-before the
revisiting write, but also events that are porf-before the
revisiting write and their rf/mo-pos is predetermined by some
other event in the pporf-prefix of the write.

The checks for co/rb edges from and toward the revisiting
prefix are modified accordingly.

TODO:
    - The same (or similar) changes will have to be migrated to
      IMM-WB

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>