execgraph*,driver: Remove inconsistent revisits before coh checks
When the graph is an inconsistent state e.g.:

  Rx0 Rx1 Rx0
  Wx1 Wx2 Wx1

(each thread contains an RMW instruction) we can only revisit
reads that were added before the conflicting read (i.e., we cannot
revisit Rx1 in this case).

Instead of filtering out such options after checking coherence,
do so before so that we avoid costly calculations in the wb
case. (It also looks nicer.)

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>